backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
langchain==0.3.7
langchain-google-genai==2.0.5
langchain-community==0.3.5
chromadb==0.5.20
sentence-transformers==3.3.1
huggingface-hub==0.26.2
